requirements.txt
onnxruntime-gpu==1.15.1
numpy
opencv-python
pillow
tqdm